requirements.txt
Django>=5.1,<6.0
djangorestframework
python-dotenv
mysqlclient
